Move DatabaseOperations to a more appropriate place.
